refactor: attach images action is now on relation manager header
