+++ app/Filament/Resources/Wiki/Anime/Pages/EditAnime.php
@@ -8,7 +8,6 @@ use App\Enums\Models\Wiki\ResourceSite;
 use App\Filament\HeaderActions\Discord\DiscordThreadHeaderAction;
 use App\Filament\HeaderActions\Models\Wiki\Anime\AttachAnimeResourceHeaderAction;
 use App\Filament\HeaderActions\Models\Wiki\Anime\BackfillAnimeHeaderAction;
-use App\Filament\HeaderActions\Models\Wiki\AttachImageHeaderAction;
 use App\Filament\Resources\Wiki\Anime;
 use App\Filament\Resources\Base\BaseEditResource;
 use Filament\Actions\ActionGroup;
@@ -45,8 +44,6 @@ class EditAnime extends BaseEditResource
                     DiscordThreadHeaderAction::make('discord-thread-header'),
 
                     BackfillAnimeHeaderAction::make('backfill-anime'),
-
-                    AttachImageHeaderAction::make('attach-anime-image'),
 
                     AttachAnimeResourceHeaderAction::make('attach-anime-resource'),
