+++ app/Filament/Resources/Wiki/Studio/Pages/EditStudio.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Filament\Resources\Wiki\Studio\Pages;
 
-use App\Filament\HeaderActions\Models\Wiki\AttachImageHeaderAction;
 use App\Filament\HeaderActions\Models\Wiki\Studio\AttachStudioResourceHeaderAction;
 use App\Filament\HeaderActions\Models\Wiki\Studio\BackfillStudioHeaderAction;
 use App\Filament\Resources\Wiki\Studio;
@@ -31,8 +30,6 @@ class EditStudio extends BaseEditResource
             [
                 ActionGroup::make([
                     BackfillStudioHeaderAction::make('backfill-studio'),
-
-                    AttachImageHeaderAction::make('attach-studio-image'),
 
                     AttachStudioResourceHeaderAction::make('attach-studio-resource'),
                 ])
